+++ samples/elaborate/tetris/view/block.rb
@@ -24,7 +24,7 @@ class Tetris
     class Block
       include Glimmer::UI::CustomWidget
   
-      options :block_size, :row, :column
+      options :game_playfield, :block_size, :row, :column
   
       body {
         composite {
@@ -33,7 +33,7 @@ class Tetris
             width_hint block_size
             height_hint block_size
           }
-          background bind(Model::Game.playfield[row][column], :color)
+          background bind(game_playfield[row][column], :color)
           # TODO improve shapes to have a bevel look
           rectangle(0, 0, block_size, block_size)
           rectangle(3, 3, block_size - 6, block_size - 6) {